renames LinkData to LinkedData
